Move create data output to other API
