+++ backend/app/data_outputs/router.py
@@ -7,12 +7,10 @@ from app.core.auth.auth import get_authenticated_user
 from app.core.authz.actions import AuthorizationAction
 from app.core.authz.authorization import (
     Authorization,
-    DataOutputOwnerResolver,
     DataOutputResolver,
 )
 from app.data_outputs.schema import (
     DataOutput,
-    DataOutputCreate,
     DataOutputStatusUpdate,
     DataOutputUpdate,
 )
@@ -63,36 +61,6 @@ def remove_data_output(
     return DataOutputService().remove_data_output(id, db, authenticated_user)
 
 
-@router.post(
-    "",
-    responses={
-        200: {
-            "description": "DataOutput successfully created",
-            "content": {
-                "application/json": {
-                    "example": {"id": "random id of the new data_output"}
-                }
-            },
-        },
-    },
-    dependencies=[
-        Depends(
-            Authorization.enforce(
-                AuthorizationAction.DATA_PRODUCT__CREATE_DATA_OUTPUT,
-                DataOutputOwnerResolver,
-                object_id="owner_id",
-            )
-        )
-    ],
-)
-def create_data_output(
-    data_output: DataOutputCreate,
-    db: Session = Depends(get_db_session),
-    authenticated_user: User = Depends(get_authenticated_user),
-) -> dict[str, UUID]:
-    return DataOutputService().create_data_output(data_output, db, authenticated_user)
-
-
 @router.put(
     "/{id}",
     responses={
@@ -104,7 +72,7 @@ def create_data_output(
         }
     },
     dependencies=[
-        Depends(only_data_output_owners),
+        # Depends(only_data_output_owners),
         Depends(
             Authorization.enforce(
                 AuthorizationAction.DATA_PRODUCT__UPDATE_DATA_OUTPUT,
