+++ backend/app/data_outputs/schema.py
@@ -8,17 +8,21 @@ from app.shared.schema import ORMModel
 from app.tags.schema import Tag
 
 
-class DataOutputCreate(ORMModel):
+class DataOutputCreateRequest(ORMModel):
     name: str
     description: str
     external_id: str
-    owner_id: UUID
+
     platform_id: UUID
     service_id: UUID
     status: DataOutputStatus
     configuration: DataOutputConfiguration
     sourceAligned: bool
     tag_ids: list[UUID]
+
+
+class DataOutputCreate(DataOutputCreateRequest):
+    owner_id: UUID
 
 
 class DataOutputUpdate(ORMModel):
