+++ backend/app/data_products/router.py
@@ -6,7 +6,9 @@ from sqlalchemy.orm import Session
 from app.core.auth.auth import get_authenticated_user
 from app.core.authz.actions import AuthorizationAction
 from app.core.authz.authorization import Authorization, DataProductResolver
+from app.data_outputs.schema import DataOutputCreateRequest
 from app.data_outputs.schema_get import DataOutputGet
+from app.data_outputs.service import DataOutputService
 from app.data_product_memberships.enums import DataProductUserRole
 from app.data_product_settings.service import DataProductSettingService
 from app.data_products.schema import (
@@ -115,6 +117,39 @@ def update_data_product(
     id: UUID, data_product: DataProductUpdate, db: Session = Depends(get_db_session)
 ):
     return DataProductService().update_data_product(id, data_product, db)
+
+
+@router.post(
+    "/{id}/data_output",
+    responses={
+        200: {
+            "description": "DataOutput successfully created",
+            "content": {
+                "application/json": {
+                    "example": {"id": "random id of the new data_output"}
+                }
+            },
+        },
+    },
+    dependencies=[
+        Depends(
+            Authorization.enforce(
+                AuthorizationAction.DATA_PRODUCT__CREATE_DATA_OUTPUT,
+                DataProductResolver,
+                object_id="owner_id",
+            )
+        )
+    ],
+)
+def create_data_output(
+    id: UUID,
+    data_output: DataOutputCreateRequest,
+    db: Session = Depends(get_db_session),
+    authenticated_user: User = Depends(get_authenticated_user),
+) -> dict[str, UUID]:
+    return DataOutputService().create_data_output(
+        id, data_output, db, authenticated_user
+    )
 
 
 @router.put(
