+++ frontend/src/api/api-urls.ts
@@ -17,6 +17,7 @@ export enum ApiUrl {
     DataProductLifecycle = '/api/data_product_lifecycles',
     DataProductLifecycleGet = '/api/data_product_lifecycles/:lifecycleId',
     DataProductSetting = '/api/data_product_settings',
+    DataProductOutputCreate = '/api/data_products/:dataProductId/data_output',
     DataProductSettingGet = '/api/data_product_settings/:settingId',
     DataProductSettingValue = '/api/data_products/:dataProductId/settings/:dataProductSettingId',
     DataProductSignInUrl = '/api/data_products/:dataProductId/signin_url',
