+++ frontend/src/types/data-output/data-output-create.contract.ts
@@ -4,7 +4,6 @@ export type DataOutputCreate = Pick<
     DataOutputContract,
     'name' | 'status' | 'configuration' | 'description' | 'external_id' | 'tag_ids'
 > & {
-    owner_id: string;
     platform_id: string;
     service_id: string;
     sourceAligned: boolean;
